Move TestPerformance to a new 'benchmark' folder
This commit is a minimal implementation of work previously done by
@BurtHarris in light of other changes made to the master branch build
strategy.
