+++ .gitignore
@@ -39,6 +39,7 @@ jspm_packages/
 # Build products
 /src/tree/xpath/XPathLexer.tokens
 /src/tree/xpath/XPathLexer.ts
+/benchmark/gen/
 /test/tool/gen/
 /target/
 /tool/target/
