+++ package.json
@@ -11,19 +11,27 @@
     "buildtool": "cd tool && npm link",
     "unlinktool": "cd tool && npm unlink",
     "clean": "npm run unlink && git clean -idx",
-    "antlr4ts": "npm run antlr4ts-xpath && npm run antlr4ts-test-std && npm run antlr4ts-test-std-atn && npm run antlr4ts-test-lr && npm run antlr4ts-test-lr-atn",
+    "antlr4ts": "npm run antlr4ts-xpath && npm run antlr4ts-benchmark",
     "antlr4ts-xpath": "antlr4ts src/tree/xpath/XPathLexer.g4 -DbaseImportPath=../.. -o src/tree/xpath",
-    "antlr4ts-test-std": "antlr4ts -visitor test/tool/Java.g4 -DbaseImportPath=../../../../src -o test/tool/gen/std",
-    "antlr4ts-test-std-atn": "antlr4ts -visitor -Xforce-atn test/tool/Java.g4 -DbaseImportPath=../../../../src -o test/tool/gen/std-atn",
-    "antlr4ts-test-lr": "antlr4ts -visitor test/tool/JavaLR.g4 -DbaseImportPath=../../../../src -o test/tool/gen/lr",
-    "antlr4ts-test-lr-atn": "antlr4ts -visitor -Xforce-atn test/tool/JavaLR.g4 -DbaseImportPath=../../../../src -o test/tool/gen/lr-atn",
-    "tsc": "tsc","test": "mocha",
+    "antlr4ts-benchmark": "npm run antlr4ts-benchmark-std && npm run antlr4ts-benchmark-std-atn && npm run antlr4ts-benchmark-lr && npm run antlr4ts-benchmark-lr-atn",
+    "antlr4ts-benchmark-std": "antlr4ts -visitor benchmark/Java.g4 -DbaseImportPath=../../../src -o benchmark/gen/std",
+    "antlr4ts-benchmark-std-atn": "antlr4ts -visitor -Xforce-atn benchmark/Java.g4 -DbaseImportPath=../../../src -o benchmark/gen/std-atn",
+    "antlr4ts-benchmark-lr": "antlr4ts -visitor benchmark/JavaLR.g4 -DbaseImportPath=../../../src -o benchmark/gen/lr",
+    "antlr4ts-benchmark-lr-atn": "antlr4ts -visitor -Xforce-atn benchmark/JavaLR.g4 -DbaseImportPath=../../../src -o benchmark/gen/lr-atn",
+    "tsc": "tsc",
+    "test": "mocha",
     "cover": "istanbul cover node_modules/mocha/bin/_mocha",
     "postcover": "istanbul report",
     "dev-test-watch": "mocha-typescript-watch",
     "typedoc": "typedoc --ignoreCompilerErrors --out doc src --name antlr4ts",
     "pack-runtime": "npm pack target/src",
-    "publish-runtime": "npm install && npm publish target/src"
+    "publish-runtime": "npm install && npm publish target/src",
+    "prebenchmark": "npm run antlr4ts-benchmark && tsc",
+    "benchmark": "npm run benchmark-core",
+    "benchmark-core": "node target/benchmark/TestPerformance.js",
+    "preprofile": "npm run antlr4ts-benchmark && tsc",
+    "profile": "npm run profile-core",
+    "profile-core": "node --prof target/benchmark/TestPerformance.js"
   },
   "repository": {
     "type": "git",
